Add files to convert map images to polylines
The code in this subdirectory is responsible for "guided vectorization", our method for extracting linear features from raster images of road maps that we get from API queries. The directory also includes provider-specific scripts that use ImageMagick to convert color images to binary images where roads are white and everything else is black.